+++ src/main/java/net/rptools/maptool/client/ui/zone/renderer/ZoneRenderer.java
@@ -85,9 +85,7 @@ public class ZoneRenderer extends JComponent
 
   private static final long serialVersionUID = 3832897780066104884L;
   private static final Logger log = LogManager.getLogger(ZoneRenderer.class);
-
-  private static final Color TRANSLUCENT_YELLOW =
-      new Color(Color.yellow.getRed(), Color.yellow.getGreen(), Color.yellow.getBlue(), 50);
+  private static final ZoneRendererConstants constants = new ZoneRendererConstants();
 
   /** DebounceExecutor for throttling repaint() requests. */
   private final DebounceExecutor repaintDebouncer;
@@ -98,7 +96,6 @@ public class ZoneRenderer extends JComponent
   /** Is the noise filter on for disrupting pattens in background tiled textures. */
   private boolean bgTextureNoiseFilterOn = false;
 
-  public static final int MIN_GRID_SIZE = 10;
   private static LightSourceIconOverlay lightSourceIconOverlay = new LightSourceIconOverlay();
 
   /** The zone the ZoneRenderer was built from. */
@@ -158,6 +155,7 @@ public class ZoneRenderer extends JComponent
   private ZonePoint previousZonePoint;
 
   private final EnumSet<Layer> disabledLayers = EnumSet.noneOf(Layer.class);
+  private final ZoneCompositor compositor;
 
   /**
    * Constructor for the ZoneRenderer from a zone.
@@ -169,7 +167,7 @@ public class ZoneRenderer extends JComponent
       throw new IllegalArgumentException("Zone cannot be null");
     }
     this.zone = zone;
-
+    this.compositor = new ZoneCompositor();
     repaintDebouncer = new DebounceExecutor(1000 / AppPreferences.getFrameRateCap(), this::repaint);
 
     setFocusable(true);
@@ -882,126 +880,8 @@ public class ZoneRenderer extends JComponent
     return new PlayerView(role, selectedTokens);
   }
 
-  public Rectangle fogExtents() {
-    return zone.getExposedArea().getBounds();
-  }
 
-  /**
-   * Get a bounding box, in Zone coordinates, of all the elements in the zone. This method was
-   * created by copying renderZone() and then replacing each bit of rendering with a routine to
-   * simply aggregate the extents of the object that would have been rendered.
-   *
-   * @param view the player view
-   * @return a new Rectangle with the bounding box of all the elements in the Zone
-   */
-  public Rectangle zoneExtents(PlayerView view) {
-    // Can't initialize extents to any set x/y values, because
-    // we don't know if the actual map contains that x/y.
-    // So we need a flag to say extents is 'unset', and the best I
-    // could come up with is checking for 'null' on each loop iteration.
-    Rectangle extents = null;
 
-    // We don't iterate over the layers in the same order as rendering
-    // because its cleaner to group them by type and the order doesn't matter.
-
-    // First background image extents
-    // TODO: when the background image can be resized, fix this!
-    if (zone.getMapAssetId() != null) {
-      extents =
-          new Rectangle(
-              zone.getBoardX(),
-              zone.getBoardY(),
-              ImageManager.getImage(zone.getMapAssetId(), this).getWidth(),
-              ImageManager.getImage(zone.getMapAssetId(), this).getHeight());
-    }
-    // next, extents of drawing objects
-    List<DrawnElement> drawableList = zone.getAllDrawnElements();
-    for (DrawnElement element : drawableList) {
-      if (!view.isGMView() && !element.getDrawable().getLayer().isVisibleToPlayers()) {
-        continue;
-      }
-
-      Drawable drawable = element.getDrawable();
-      Rectangle drawnBounds = new Rectangle(drawable.getBounds());
-
-      // Handle pen size
-      // This slightly over-estimates the size of the pen, but we want to
-      // make sure to include the anti-aliased edges.
-      Pen pen = element.getPen();
-      int penSize = (int) Math.ceil((pen.getThickness() / 2) + 1);
-      drawnBounds.setBounds(
-          drawnBounds.x - penSize,
-          drawnBounds.y - penSize,
-          drawnBounds.width + (penSize * 2),
-          drawnBounds.height + (penSize * 2));
-
-      if (extents == null) {
-        extents = drawnBounds;
-      } else {
-        extents.add(drawnBounds);
-      }
-    }
-    // now, add the stamps/tokens
-    // tokens and stamps are the same thing, just treated differently
-
-    // Note: order doesn't matter, so don't need to go back-to-front.
-    for (Token element :
-        zone.getTokensForLayers(layer -> view.isGMView() || layer.isVisibleToPlayers())) {
-      Rectangle drawnBounds = element.getBounds(zone);
-      if (element.hasFacing()) {
-        // Get the facing and do a quick fix to make the math easier: -90 is 'unrotated' for some
-        // reason
-        int facing = element.getFacing() + 90;
-        if (facing > 180) {
-          facing -= 360;
-        }
-        // if 90 degrees, just swap w and h
-        // also swap them if rotated more than 90 (optimization for non-90deg rotations)
-        if (facing != 0 && facing != 180) {
-          if (Math.abs(facing) >= 90) {
-            drawnBounds.setSize(drawnBounds.height, drawnBounds.width); // swapping h and w
-          }
-          // if rotated to non-axis direction, assume the worst case 45 deg
-          // also assumes the rectangle rotates around its center
-          // This will usually makes the bounds bigger than necessary, but its quick.
-          // Also, for quickness, we assume its a square token using the larger dimension
-          // At 45 deg, the bounds of the square will be sqrt(2) bigger, and the UL corner will
-          // shift by 1/2 of the length.
-          // The size increase is: (sqrt*(2) - 1) * size ~= 0.42 * size.
-          if (facing != 0 && facing != 180 && facing != 90 && facing != -90) {
-            int size = Math.max(drawnBounds.width, drawnBounds.height);
-            int x = drawnBounds.x - (int) (0.21 * size);
-            int y = drawnBounds.y - (int) (0.21 * size);
-            int w = drawnBounds.width + (int) (0.42 * size);
-            int h = drawnBounds.height + (int) (0.42 * size);
-            drawnBounds.setBounds(x, y, w, h);
-          }
-        }
-      }
-      // TODO: Handle auras here?
-      if (extents == null) {
-        extents = drawnBounds;
-      } else {
-        extents.add(drawnBounds);
-      }
-    }
-    if (zone.hasFog()) {
-      if (extents == null) {
-        extents = fogExtents();
-      } else {
-        extents.add(fogExtents());
-      }
-    }
-    // TODO: What are token templates?
-    // renderTokenTemplates(g2d, view);
-
-    // TODO: Do lights make the area of interest larger?
-    // see: renderLights(g2d, view);
-
-    // TODO: Do auras make the area of interest larger?
-    // see: renderAuras(g2d, view);
-    return extents;
-  }
 
   /**
    * This method clears {@link #visibleScreenArea} and {@link #lastView}. It also flushes the {@link
@@ -1034,10 +914,21 @@ public class ZoneRenderer extends JComponent
    */
   public void renderZone(Graphics2D g2d, PlayerView view) {
     timer.start("setup");
+    // store previous rendering settings
+    RenderingHints oldRenderingHints = g2d.getRenderingHints();
+
+    // Clear internal state
+    tokenLocationMap.clear();
+    markerLocationList.clear();
+    itemRenderList.clear();
+
+    if (!compositor.isInitialised()) compositor.setRenderer(this);
+
+    Rectangle viewRect = new Rectangle(getSize().width, getSize().height);
+
     g2d.setFont(AppStyle.labelFont);
     Object oldAA = SwingUtil.useAntiAliasing(g2d);
 
-    Rectangle viewRect = new Rectangle(getSize().width, getSize().height);
     Area viewArea = new Area(viewRect);
     // much of the raster code assumes the user clip is set
     boolean resetClip = false;
@@ -1067,19 +958,9 @@ public class ZoneRenderer extends JComponent
     }
     lastView = view;
 
-    // Clear internal state
-    tokenLocationMap.clear();
-    markerLocationList.clear();
-    itemRenderList.clear();
+    Map<Token, Set<Token>> drawThese = compositor.drawWhat(viewRect);
 
     timer.stop("setup");
-
-    // Calculations
-    timer.start("calcs-1");
-    AffineTransform af = new AffineTransform();
-    af.translate(zoneScale.getOffsetX(), zoneScale.getOffsetY());
-    af.scale(getScale(), getScale());
-
     // @formatter:off
     /*
      * This is the new code that doesn't work. See below for newer code that _might_ work. ;-) if (visibleScreenArea
@@ -1090,6 +971,12 @@ public class ZoneRenderer extends JComponent
      * Rectangle(0, 0, getSize().width, getSize().height)); }
      */
     // @formatter:on
+
+    // Calculations
+    timer.start("calcs-1");
+    AffineTransform af = new AffineTransform();
+    af.translate(zoneScale.getOffsetX(), zoneScale.getOffsetY());
+    af.scale(getScale(), getScale());
 
     if (visibleScreenArea == null) {
       timer.start("ZoneRenderer-getVisibleArea");
@@ -1354,7 +1241,8 @@ public class ZoneRenderer extends JComponent
     // zoneScale.getOffsetY()));
     // g2d.draw(area);
     // }
-    SwingUtil.restoreAntiAliasing(g2d, oldAA);
+    g2d.setRenderingHints(oldRenderingHints);
+
     if (resetClip) {
       g2d.setClip(null);
     }
@@ -2024,7 +1912,7 @@ public class ZoneRenderer extends JComponent
 
   protected void renderGrid(Graphics2D g, PlayerView view) {
     int gridSize = (int) (zone.getGrid().getSize() * getScale());
-    if (!AppState.isShowGrid() || gridSize < MIN_GRID_SIZE) {
+    if (!AppState.isShowGrid() || gridSize < constants.MIN_GRID_SIZE) {
       return;
     }
     zone.getGrid().draw(this, g, g.getClipBounds());
@@ -3302,7 +3190,7 @@ public class ZoneRenderer extends JComponent
             if (token.getFacing() < 0) {
               tokenG.setColor(Color.yellow);
             } else {
-              tokenG.setColor(TRANSLUCENT_YELLOW);
+              tokenG.setColor(constants.TRANSLUCENT_YELLOW);
             }
             tokenG.fill(arrow);
             tokenG.setColor(Color.darkGray);
